move general datatypes to g.py
