+++ src/climatevision/generator/heat2030/energy_general/g.py
@@ -1,0 +1,49 @@
+# pyright: strict
+
+from dataclasses import dataclass
+
+
+@dataclass(kw_only=True)
+class Vars1:
+    # Used by g
+    cost_wage: float = None  # type: ignore
+    demand_emplo: float = None  # type: ignore
+    demand_emplo_com: float = None  # type: ignore
+    demand_emplo_new: float = None  # type: ignore
+    invest: float = None  # type: ignore
+    invest_com: float = None  # type: ignore
+    invest_pa: float = None  # type: ignore
+    invest_pa_com: float = None  # type: ignore
+
+
+@dataclass(kw_only=True)
+class Vars2:
+    # Used by g_storage
+    cost_wage: float = None  # type: ignore
+    demand_emplo: float = None  # type: ignore
+    demand_emplo_new: float = None  # type: ignore
+    energy: float = None  # type: ignore
+    invest: float = None  # type: ignore
+    invest_com: float = None  # type: ignore
+    invest_pa: float = None  # type: ignore
+    invest_pa_com: float = None  # type: ignore
+    invest_per_x: float = None  # type: ignore
+    pct_energy: float = None  # type: ignore
+    pct_of_wage: float = None  # type: ignore
+    power_to_be_installed: float = None  # type: ignore
+    ratio_wage_to_emplo: float = None  # type: ignore
+
+
+@dataclass(kw_only=True)
+class Vars3:
+    # Used by g_planning
+    cost_wage: float = None  # type: ignore
+    demand_emplo: float = None  # type: ignore
+    demand_emplo_com: float = None  # type: ignore
+    demand_emplo_new: float = None  # type: ignore
+    invest: float = None  # type: ignore
+    invest_com: float = None  # type: ignore
+    invest_pa: float = None  # type: ignore
+    invest_pa_com: float = None  # type: ignore
+    pct_of_wage: float = None  # type: ignore
+    ratio_wage_to_emplo: float = None  # type: ignore
